Removing conversion project controllers & backing code. This has now moved to the academisation API.
