--- TramsDataApi/ResponseModels/AcademyConversionProject/AcademyConversionProjectNoteResponse.cs
@@ -1,12 +1,0 @@
-using System;
-
-namespace TramsDataApi.ResponseModels.AcademyConversionProject
-{
-    public class AcademyConversionProjectNoteResponse
-    {
-        public string Subject { get; set; }
-        public string Note { get; set; }
-        public string Author { get; set; }
-        public DateTime Date { get; set; }
-    }
-}
