--- TramsDataApi/UseCases/IGetAcademyConversionProject.cs
@@ -1,10 +1,0 @@
-using System.Threading.Tasks;
-using TramsDataApi.ResponseModels.AcademyConversionProject;
-
-namespace TramsDataApi.UseCases
-{
-    public interface IGetAcademyConversionProject
-    {
-        Task<AcademyConversionProjectResponse> Execute(int id);
-    }
-}
