--- TramsDataApi/UseCases/IGetAcademyConversionProjectFilterParameters.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-using TramsDataApi.ResponseModels;
-
-namespace TramsDataApi.UseCases
-{
-	public interface IGetAcademyConversionProjectFilterParameters
-    {
-        Task<ProjectFilterParameters> Execute();
-    }
-}
