+++ TramsDataApi/UseCases/SearchAcademyTransferProjects.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using TramsDataApi.DatabaseModels;
 using TramsDataApi.Gateways;
+using TramsDataApi.RequestModels;
 using TramsDataApi.ResponseModels;
 using TramsDataApi.ResponseModels.AcademyTransferProject;
